create attachment wrapper services for moderator, manager and auditor

+++ attachment/service_moderator.py
@@ -1,6 +1,7 @@
 
 from kronos.exceptions import ObjectNotFound
 
+from audit_store.models import AuditStore
 import audit_store.service_moderator as audit_store_service
 
 from .models import Attachment
@@ -13,6 +14,10 @@ def find_by_audit_store_for_moderator(audit_store_id, user_id):
 
 def upload_for_audit_store_for_moderator(audit_store_id, user_id, file_name, file_size, mime_type):
     audit_store = audit_store_service.find_by_id_for_moderator(audit_store_id, user_id)
+
+    if audit_store.status != AuditStore.SUBMITTED:
+        raise AppLogicError("cannot upload attachment now")
+
     return attachment_service.upload_for_audit_store(audit_store.id, file_name, file_size, mime_type)
 
 def rename_for_moderator(attachment_id, user_id, new_name):
@@ -23,9 +28,17 @@ def rename_for_moderator(attachment_id, user_id, new_name):
 def complete_for_moderator(attachment_id, user_id):
     audit_store = attachment_service.get_audit_store_for_attachment(attachment_id)
     audit_store_service.find_by_id_for_moderator(audit_store.id, user_id)
+
+    if audit_store.status != AuditStore.SUBMITTED:
+        raise AppLogicError("cannot complete attachment now")
+
     return attachment_service.complete(attachment_id)
 
 def delete_for_moderator(attachment_id, user_id):
     audit_store = attachment_service.get_audit_store_for_attachment(attachment_id)
     audit_store_service.find_by_id_for_moderator(audit_store.id, user_id)
+
+    if audit_store.status != AuditStore.SUBMITTED:
+        raise AppLogicError("cannot delete attachment now")
+
     return attachment_service.delete(attachment_id)
